dollclaw map update
